add type dynamic in courrier_deprt and arrivee and add tcpdf

+++ database/migrations/2023_02_28_121546_create_courrier_arrives_table.php
@@ -33,7 +33,7 @@ return new class extends Migration
             $table->string('etudiant');
             $table->integer('is_rep');
             $table->dateTime('date_rep');
-            $table->integer('etat_courrier_id');
+            $table->unsignedBigInteger('etat_courrier_id');
             $table->integer('is_lu');
             $table->timestamps();
         });
